julia: fix array indexing logic
Fixes https://github.com/adsharma/py2many/issues/88

+++ tests/expected/binit.rs
@@ -5,7 +5,7 @@ fn bisect_right(data: &Vec<i32>, item: i32) -> i32 {
     let mut high: _ = data.len();
     while low < high {
         let middle: _ = i32::from(((low + high) / 2));
-        if item < data[middle] {
+        if item < data[middle as usize] {
             high = middle;
         } else {
             low = (middle + 1);
